Go back to an incremental hart_request API (with a twist)
Requests can now be negative.  The absolute interface had the problem
pointed out a few commits ago, and putting a lock around the problem
made things slow down significantly. The absolute API was what we want
for performance, but the relative API is what we need for correctness.
This solution finds the right middle ground to make this work in all
cases.

The problem with the absolute API is that passing an absolute count up
the request chain can get out of sync with the original count due to
races when running on multiple cores.  The count maintained by any
individual scheduler first gets incremented, then passed up the
request chain as an absolute value.  If one core is slower than
another, the final value might be updated out-of-order from the
original value set at the beginning of the call chain (unless the
entire chain is protected by a lock). For example, the original count
might be updated from 1 to 0 in the proper order, but the final value
could end up as 1, when 0 was the desired value.  Passing an
incremental value up the call chain doesn't have this problem, because
the final value isn't adjusted until the very end. It will always
reach the desired value eventually.

+++ src/defaults.c
@@ -34,7 +34,7 @@ void __lithe_context_destroy_default(lithe_context_t *context, bool stack)
 }
 
 /* Helpful defaults for the lithe callback functions */
-void __hart_request_default(lithe_sched_t *__this, lithe_sched_t *child, size_t h)
+void __hart_request_default(lithe_sched_t *__this, lithe_sched_t *child, int h)
 {
   fatal("You need to define the hart_request callback yourself!");
 }
