+++ src/lithe.h
@@ -57,7 +57,7 @@ lithe_sched_t *lithe_sched_current();
  * the requesting scheduler. Returns the number of harts successfully
  * requested.
  */
-void lithe_hart_request(size_t h);
+void lithe_hart_request(int h);
  
 /**
  * Grant the current hart to another scheduler.  Triggered by a previous call
